Tim 9/requirements.txt
cachetools==5.2.0
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
entrypoints==0.4
Flask==2.1.2
flatbuffers==1.12
gast==0.4.0
gevent==21.12.0
greenlet==1.1.2
grpcio==1.46.3
h5py==3.7.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
mistune==0.8.4
nest-asyncio==1.5.5
numpy==1.22.4
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
pickleshare==0.7.5
portpicker==1.2.0
prompt-toolkit==1.0.18
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
requests==2.21.0
soupsieve==2.3.2.post1
urllib3==1.24.3
webencodings==0.5.1
Werkzeug==2.1.2
tensorflow-cpu==2.8.0
gunicorn
pillow>=9.1.0
matplotlib